Move yoyo code to own submodule
This is just to get the code outside of adcp where it doesn't belong. In
the future it will hopefully be part of the towyo submodule.
